Convert bootstrapper tests to mocha
